+++ Modules/Networking/networking_http_server.cpp
@@ -611,7 +611,7 @@ int32_t ft_http_server::run_once_locked()
         if (is_post && !body.empty())
         {
             std::snprintf(buffer, sizeof(buffer), "%" PRIu64,
-                static_cast<uint64_t>(body.size()));
+                body.size());
             response.append("HTTP/1.1 200 OK\r\nContent-Length: ");
             response.append(buffer);
             if (should_keep_alive != FT_FALSE)
