+++ Modules/Observability/observability.cpp
@@ -94,9 +94,9 @@ int32_t observability_record_operation(ft_observability_module module,
         storage->successes.fetch_add(1, std::memory_order_relaxed);
     else
         storage->failures.fetch_add(1, std::memory_order_relaxed);
-    storage->bytes_read.fetch_add(static_cast<uint64_t>(bytes_read),
+    storage->bytes_read.fetch_add(bytes_read,
         std::memory_order_relaxed);
-    storage->bytes_written.fetch_add(static_cast<uint64_t>(bytes_written),
+    storage->bytes_written.fetch_add(bytes_written,
         std::memory_order_relaxed);
     return (FT_ERR_SUCCESS);
 }
